python/requirements.txt
# Python библиотеки для генерации документов

# Работа с DOCX файлами
python-docx==1.1.0

# Генерация PDF
reportlab==4.0.7

# Шаблонизация
jinja2==3.1.2

# Обработка изображений
Pillow==10.2.0

# Работа с Excel (опционально)
openpyxl==3.1.2

# Утилиты
python-dateutil==2.8.2

